Use int type for command line lengths.

+++ src/common/cmd.c
@@ -774,10 +774,10 @@ static  char    *cmd_null_string = "";
 
 // complete command string, left untouched
 static  char    cmd_string[MAX_STRING_CHARS];
-static  size_t  cmd_string_len;
+static  int     cmd_string_len;
 
 // offsets of individual tokens into cmd_string
-static  size_t  cmd_offsets[MAX_STRING_TOKENS];
+static  int     cmd_offsets[MAX_STRING_TOKENS];
 
 // sequence of NULL-terminated, normalized tokens
 static  char    cmd_data[MAX_STRING_CHARS];
@@ -794,7 +794,7 @@ from_t Cmd_From(void)
     return cmd_current->from;
 }
 
-size_t Cmd_ArgOffset(int arg)
+int Cmd_ArgOffset(int arg)
 {
     if (arg < 0) {
         return 0;
@@ -805,7 +805,7 @@ size_t Cmd_ArgOffset(int arg)
     return cmd_offsets[arg];
 }
 
-int Cmd_FindArgForOffset(size_t offset)
+int Cmd_FindArgForOffset(int offset)
 {
     int i;
